Page describing Haskell.org and the Haskell.org committee
This adds an "official" page describing the Haskell.org non-profit and the Haskell.org committee. The content is based on the [Haskell.org committee wiki page][wiki], edited a bit for focus and clarity.

Over time, I expect this page will cover more details about how the committee operates and makes decisions—this is the first step to making the committee's activities and structure more transparent to the community.

[wiki]: https://wiki.haskell.org/Haskell.org_committee